Removed UIToolkit asset from InputReader control
Disabling the UI Toolkit input action asset right in the
middle of a user click will force the system to miss the
up-click of the mouse and will require a dummy click upon re-enable to match it with a corresponding down-click before being able to focus on an element again. There may be a better way of doing this, but for the sake of time, it will be removed.